% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_details.R
\name{createStrategusConnectionDetails}
\alias{createStrategusConnectionDetails}
\title{Create DatabaseConnector connectionDetails from strategus-db-details.json}
\usage{
createStrategusConnectionDetails(
  path = file.path(getwd(), "strategus-db-details.json"),
  dbDetails = NULL
)
}
\arguments{
\item{path}{path to strategus-db-details.json}

\item{dbDetails}{optional list of db settings (if already loaded)}
}
\value{
DatabaseConnector connectionDetails object
}
\description{
Create DatabaseConnector connectionDetails from strategus-db-details.json
}
