% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_details.R
\name{readStrategusDbDetails}
\alias{readStrategusDbDetails}
\title{Read Strategus database details from JSON}
\usage{
readStrategusDbDetails(path = file.path(getwd(), "strategus-db-details.json"))
}
\arguments{
\item{path}{path to strategus-db-details.json}
}
\value{
list of db settings
}
\description{
Read Strategus database details from JSON
}
