% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution_settings.R
\name{readStrategusExecutionSettings}
\alias{readStrategusExecutionSettings}
\title{Read Strategus execution settings from JSON}
\usage{
readStrategusExecutionSettings(
  path = file.path(getwd(), "strategus-execution-settings.json")
)
}
\arguments{
\item{path}{path to strategus-execution-settings.json}
}
\value{
list of execution settings
}
\description{
Read Strategus execution settings from JSON
}
