% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_dialogue.R
\name{workflow_dialogue_prompt_width}
\alias{workflow_dialogue_prompt_width}
\title{Render a workflow dialogue response in the interactive shell}
\usage{
workflow_dialogue_prompt_width()
}
\arguments{
\item{response}{ACP workflow dialogue response}
}
\value{
invisible NULL
}
\description{
Render a workflow dialogue response in the interactive shell
}
